fixed heap and counting sort, updated tests

+++ sorting/counting_sort.go
@@ -3,7 +3,8 @@
 */
 package sort
 
-func getK(a []int) int {
+
+func getK(a []uint) int {
 	if len(a) == 0 {
 		return 1
 	}
@@ -13,22 +14,22 @@ func getK(a []int) int {
 			k = v
 		}
 	}
-	return k+1
+	return int(k+1)
 }
 
-func CountingSort(array []int) {
+func CountingSort(array []uint) {
 	l := len(array)
 	k := getK(array)
-	b := make([]int, k)	// array of counts
+	b := make([]uint, k)	// array of counts
 	for i:= 0; i < l; i++ {
-		b[array[i]] += 1
+		b[array[i]] += uint(1)
 	}
 	for i, j := 0, 0; i < k; i++ {
 		for {
 			if b[i] > 0 {
-				array[j] = i
+				array[j] = uint(i)
 				j += 1
-				b[i] -= 1
+				b[i] -= uint(1)
 				continue
 			}
 			break
